+++ packages/common/src/api/tan-query/useCollection.ts
@@ -4,13 +4,10 @@ import { useDispatch, useSelector } from 'react-redux'
 import { userCollectionMetadataFromSDK } from '~/adapters/collection'
 import { useAppContext } from '~/context/appContext'
 import { Id, ID, OptionalId } from '~/models/Identifiers'
-import { Kind } from '~/models/Kind'
 import { getUserId } from '~/store/account/selectors'
-import { addEntries } from '~/store/cache/actions'
-import { EntriesByKind } from '~/store/cache/types'
-import { encodeHashId } from '~/utils/hashIds'
 
 import { QUERY_KEYS } from './queryKeys'
+import { primeCollectionData } from './utils/primeCollectionData'
 
 type Config = {
   staleTime?: number
@@ -35,48 +32,16 @@ export const useCollection = (collectionId: ID, config?: Config) => {
       const collection = userCollectionMetadataFromSDK(data[0])
 
       if (collection) {
-        // Prime user data from collection owner
-        if (collection.user) {
+        // Prime related entities
+        primeCollectionData({ collection, queryClient, dispatch })
+
+        // Prime collectionByPermalink cache if we have a permalink
+        if (collection.permalink) {
           queryClient.setQueryData(
-            [QUERY_KEYS.user, collection.user.user_id],
-            collection.user
+            [QUERY_KEYS.collectionByPermalink, collection.permalink],
+            collection
           )
         }
-
-        // Prime track and user data from tracks in collection
-        const entries: EntriesByKind = {
-          [Kind.COLLECTIONS]: {
-            [collection.playlist_id]: collection
-          }
-        }
-
-        if (collection.user) {
-          if (!entries[Kind.USERS]) entries[Kind.USERS] = {}
-          entries[Kind.USERS][collection.user.user_id] = collection.user
-        }
-
-        // Track and user data from tracks in collection
-        collection.tracks?.forEach((track) => {
-          if (track.track_id) {
-            // Prime track data
-            queryClient.setQueryData([QUERY_KEYS.track, track.track_id], track)
-            if (!entries[Kind.TRACKS]) entries[Kind.TRACKS] = {}
-            entries[Kind.TRACKS][track.track_id] = track
-
-            // Prime user data from track owner
-            if (track.user) {
-              queryClient.setQueryData(
-                [QUERY_KEYS.user, track.user.user_id],
-                track.user
-              )
-              if (!entries[Kind.USERS]) entries[Kind.USERS] = {}
-              entries[Kind.USERS][track.user.user_id] = track.user
-            }
-          }
-        })
-
-        // Sync all data to Redux in a single dispatch
-        dispatch(addEntries(entries, undefined, undefined, 'react-query'))
       }
 
       return collection
