+++ packages/common/src/api/tan-query/useCollectionByPermalink.ts
@@ -3,12 +3,11 @@ import { useDispatch, useSelector } from 'react-redux'
 
 import { userCollectionMetadataFromSDK } from '~/adapters/collection'
 import { useAppContext } from '~/context'
-import { Kind, OptionalId } from '~/models'
+import { OptionalId } from '~/models'
 import { getUserId } from '~/store/account/selectors'
-import { addEntries } from '~/store/cache/actions'
-import { EntriesByKind } from '~/store/cache/types'
 
 import { QUERY_KEYS } from './queryKeys'
+import { primeCollectionData } from './utils/primeCollectionData'
 
 export const playlistPermalinkToHandleAndSlug = (permalink: string) => {
   const splitPermalink = permalink.split('/')
@@ -50,48 +49,14 @@ export const useCollectionByPermalink = (
       const collection = userCollectionMetadataFromSDK(data[0])
 
       if (collection) {
-        // Prime user data from collection owner
-        if (collection.user) {
-          queryClient.setQueryData(
-            [QUERY_KEYS.user, collection.user.user_id],
-            collection.user
-          )
-        }
+        // Prime related entities
+        primeCollectionData({ collection, queryClient, dispatch })
 
-        // Prime track and user data from tracks in collection
-        const entries: EntriesByKind = {
-          [Kind.COLLECTIONS]: {
-            [collection.playlist_id]: collection
-          }
-        }
-
-        if (collection.user) {
-          if (!entries[Kind.USERS]) entries[Kind.USERS] = {}
-          entries[Kind.USERS][collection.user.user_id] = collection.user
-        }
-
-        // Track and user data from tracks in collection
-        collection.tracks?.forEach((track) => {
-          if (track.track_id) {
-            // Prime track data
-            queryClient.setQueryData([QUERY_KEYS.track, track.track_id], track)
-            if (!entries[Kind.TRACKS]) entries[Kind.TRACKS] = {}
-            entries[Kind.TRACKS][track.track_id] = track
-
-            // Prime user data from track owner
-            if (track.user) {
-              queryClient.setQueryData(
-                [QUERY_KEYS.user, track.user.user_id],
-                track.user
-              )
-              if (!entries[Kind.USERS]) entries[Kind.USERS] = {}
-              entries[Kind.USERS][track.user.user_id] = track.user
-            }
-          }
-        })
-
-        // Sync all data to Redux in a single dispatch
-        dispatch(addEntries(entries, undefined, undefined, 'react-query'))
+        // Prime collection cache
+        queryClient.setQueryData(
+          [QUERY_KEYS.collection, collection.playlist_id],
+          collection
+        )
       }
 
       return collection
